Check Argument's Length

Check if an argument has valid length, and if not, generate an error message.

check_length(
  x,
  valid,
  interval = NULL,
  name = NULL,
  general = NULL,
  specific = NULL,
  supplement = NULL,
  ...
)

Arguments

x
The argument to check, which can be any object.

valid
A numeric vector which contains non-negative integers and NA, used with argument interval to indicate the valid lengths.

interval
Optional. TRUE or FALSE which indicates if argument valid is interpreted as an interval or as single lengths. For example, c(1, 10) is interpreted as "larger than 1 and smaller than 10" if interval is TRUE, but as "1 or 10" if FALSE. NA can be used in valid when treated as interval. For example, c(0, NA) means "larger than 0". By default, interval is inferred from valid. For example, if valid has length unequal to 2, it's treated as single lengths.

name
A single character which gives the argument's name. The name is used in the error message. By default, the name of the argument passed to argument x is captured automatically.

general
Optional. A single character which is used to give a general statement of the error incurred. By default, this is generated automatically.

specific
Optional. A single character which gives a detailed description of the error. glue::glue() syntax can be used, see "Examples" section. By default, this is generated automatically. Can be turned off with character(0).

supplement
Optional. A (named) character vector which gives some additional information about the error. The names are used to create bullets, see Statement(). By default, this is left empty.

...
Optional. Additional arguments passed to rlang::abort(), which is called internally.

Value

An invisible NULL if the argument is valid, or an error message is generated.

See also

Examples

if (FALSE) {
x <- c(1, 2)

# `valid` as interval
check_length(x, c(1, 3), TRUE)
check_length(x, c(NA, 2))

# `valid` as single lengths
check_length(x, c(1, 3), FALSE)

# customize error message with `glue::glue()` syntax
specific <- "Oh my god! `{name}`'s length is {feature}."
check_length(x, 3, specific = specific)
}